Re-done exception handling.
Made one big exception for main part with another small one for additional methods.
