+++ FileLengthChecker/bin/Debug/netcoreapp3.1/FileLengthChecker.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v3.1",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v3.1": {
+      "FileLengthChecker/1.0.0": {
+        "runtime": {
+          "FileLengthChecker.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "FileLengthChecker/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
