+++ FileLengthChecker/bin/Debug/netcoreapp3.1/FileWeightChecker.pdb
