+++ FileLengthChecker/bin/Debug/netcoreapp3.1/FileWeightChecker.runtimeconfig.dev.json
@@ -1,0 +1,8 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\Евгений\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\Евгений\\.nuget\\packages"
+    ]
+  }
+}
